+++ todo/src/ui/components/molecules/StrictModeDroppable.tsx
