+++ todo/src/ui/components/organisms/EntityList.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import tw from "tailwind-styled-components";
-import { Board } from "./Board";
-import { List } from "./List";
+import { Board } from "../molecules/Board";
+import { List } from "../molecules/List";
 
 const Container = tw.div`
   w-full max-w-4xl mx-auto p-4
